separated spec files for text
