+++ cypress/integration/v2/blocks/text/desktop-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { desktopStyle } from '.'
+
+describe( 'Advanced Text Block ( Desktop Style )', registerTests( [ desktopStyle ] ) )
